+++ climatedata/__init__.py
@@ -1,8 +1,9 @@
 """NOAA climatedata web services interface"""
 
-__version__ = '1.0.0'
+__version__ = '1.1.0'
 __url__ = 'https://github.com/rwspicer/noaa-climatedata'
 
 from . import services
 from . import utilities
 from . import formats
+from . import datatypes