+++ climatedata/services/locationcategories.py
@@ -18,13 +18,13 @@ aliases = {
 }
 
 
-def call(**kwargs):
-    """calls service to get daily waterdata
+def call(*args, **kwargs):
+    """calls service to get location categories
 
     Parameters
     ----------
     **kwargs:
-        keyword arguments defined in waterdata.services.daily_data.filters
+        keyword arguments defined in climatedata.services.locationcategories.filters
 
     Returns
     -------
@@ -33,6 +33,14 @@ def call(**kwargs):
 
     """
     clean_args = globals.validate_args(kwargs, filters, aliases)
+
+    if len(args) == 1:
+        clean_args['specifically'] = args[0]
+    elif len(args) > 1:
+        raise generic.InvalidAPICallError(
+            'api cannot have multiple generic queries'
+    )
+    
     
     response = generic.call('locationcategories', **clean_args)