+++ climatedata/services/stations.py
@@ -14,7 +14,7 @@ filters = {
     'datacategoryid':   globals.default,
     'datatypeid':       globals.default,
     'extent':           globals.default,
-    # 'stationid':        globals.default
+    'stationid':        globals.default
 }
 
 filters.update(globals.filters)
@@ -23,13 +23,13 @@ filters.update(globals.filters)
 aliases = {
 }
 
-def call(**kwargs):
-    """calls service to get daily waterdata
+def call(*args, **kwargs):
+    """calls service to get stations
 
     Parameters
     ----------
     **kwargs:
-        keyword arguments defined in waterdata.services.daily_data.filters
+        keyword arguments defined in climatedata.services.stations.filters
 
     Returns
     -------
@@ -39,6 +39,13 @@ def call(**kwargs):
     """
     clean_args = globals.validate_args(kwargs, filters, aliases)
     
-    response = generic.call('stations', **clean_args)
+    
+    if len(args) == 1:
+        clean_args['specifically'] = args[0]
+    elif len(args) > 1:
+        raise generic.InvalidAPICallError(
+            'api cannot have multiple generic queries'
+    )
 
+    response = generic.call('stations', **clean_args)
     return json.loads(response)