+++ changelog.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## 1.1.0 [2023-02-19]
+### added
+- datatypes.py: Metadata and unit conversion functions for GHCN data types
+- added utilities.attempt function 
+- ability to call api end points without arguments to get info about a 
+specific dataset, station, etc. (does not apply to 'data' service)
+
+### changed
+- changed way utilities.depaginate works adding attempt function 
+- code documentation
+
+
 ## 1.0.0 [2022-08-25]
 ### added
 - added license